tpbase: codeql fixes (#605)
